Extract loading screen to its own package
